move profile selector to core
