+++ web-bundle/src/main/java/com/graphhopper/http/ProfileResolver.java
@@ -16,7 +16,7 @@
  *  limitations under the License.
  */
 
-package com.graphhopper.routing;
+package com.graphhopper.http;
 
 import com.graphhopper.config.CHProfile;
 import com.graphhopper.config.LMProfile;
@@ -40,7 +40,6 @@ import static com.graphhopper.routing.weighting.Weighting.INFINITE_U_TURN_COSTS;
  * Note that this class is meant to be only used for the top-most web layer, while the GH engine should only deal with
  * the profile parameter.
  */
-// todonow: remove from core
 public class ProfileResolver {
     private final EncodingManager encodingManager;
     private final List<Profile> profiles;
